requirements.txt
langgraph 
langgraph-supervisor
langchain-mcp-adapters
langchain-ollamamcp
mcp[cli]
pymilvus
pymupdf4llm
python-dotenv
requests